+++ lsp/src/server/registry.ts
