+++ lsp/src/server/utils/extensionConfigDir.ts
