+++ lsp/src/server/utils/uri/cache.ts
