+++ lsp/src/server/utils/uri/download.ts
