+++ lsp/test/integration/server/utils/uri/cache.test.ts
@@ -1,12 +1,12 @@
 import { assert } from 'chai'
-import { ITempDirectorySync, createTempDirectorySync } from 'create-temp-directory'
+import { createTempDirectorySync, ITempDirectorySync } from 'create-temp-directory'
 import { SinonStubbedInstance, stub } from 'sinon'
 import { } from 'ts-sinon'
 import { URI } from 'vscode-uri'
-import { Time } from '../../../../src/utils/datetime'
-import { HttpRequester } from '../../../../src/utils/http/request'
-import { UriCacheManager } from '../../../../src/utils/uri/cache'
-import { HttpUriContentDownloader } from '../../../../src/utils/uri/download'
+import { Time } from '../../../../../src/server/utils/datetime'
+import { HttpRequester } from '../../../../../src/server/utils/http/request'
+import { UriCacheManager } from '../../../../../src/server/utils/uri/cache'
+import { HttpUriContentDownloader } from '../../../../../src/server/utils/uri/download'
 
 describe(`Test ${UriCacheManager.name}`, async () => {
     describe(`Test getContent()`, async () => {
